🐛 Bug
jmcouffin:

* not necessarly the most elegant, should work in every languages

* missing is not element type

I think you are missing .WhereElementIsNotElementType() in your filteredeleementcollector as it is selecting both the symbol and the instance right now

ErikFrits:
I will rewrite this tool with more rules when I find some time for tihs. 
It is quite limited and does not handle well with categories like Grid, Levels. 
😉